# openapi3_1__api_gateway.yaml
openapi: 3.1.0
info:
  title: Symcon smart-home — API Gateway
  description: >-
    API шлюза сервиса управления устройствами умного дома
    для взаимодействия с клиентскими приложениями.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
servers:
  - url: https://localhost:8080/v2
    description: Test server
  - url: https://symcon.ru/v2
    description: Production server
...
paths:
...
  /groups:

    post:
      tags:
        - groups
      summary: Добавляет группу совместного пользования устройствами
      description: Добавляет группу совместного пользования устройствами
      operationId: addGroup
      security:
        - ApiTokenAuth: []
      requestBody:
        description: Данные новой группы
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '201':
          description: Группа успешно добавлена
        '400':
          description: Некорректные данные в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Пользователь не найден
        default:
          description: Ошибка сервера
  /groups/{groupNum}:

    get:
      tags:
        - groups
      summary: Ищет группу по номеру и возвращает её данные
      description: Ищет группу по номеру и возвращает её данные
      operationId: findGroupByNum
      security:
        - ApiTokenAuth: []
      parameters:
        - name: groupNum
          in: path
          description: Номер группы
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Некорректный номер группы в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Группа не найдена
        default:
          description: Ошибка сервера

    put:
      tags:
        - groups
      summary: Обновляет данные группы совместного пользования устройствами
      description: Обновляет данные группы совместного пользования устройствами
      operationId: updateGroup
      security:
        - ApiTokenAuth: []
      parameters:
        - name: groupNum
          in: path
          description: Номер группы
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Данные группы
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: Группа успешно обновлена
        '400':
          description: Некорректный номер группы в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Группа не найдена
        default:
          description: Ошибка сервера

    delete:
      tags:
        - groups
      summary: Удаляет группу совместного пользования устройствами
      description: Удаляет группу совместного пользования устройствами
      operationId: deleteGroup
      security:
        - ApiTokenAuth: []
      parameters:
        - name: groupNum
          in: path
          description: Номер группы
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Группа успешно удалена
        '400':
          description: Некорректный номер группы в запросе
        '401':
          description: Несанкционированный доступ
        '404':
          description: Группа не найдена
        default:
          description: Ошибка сервера
  /groups/{groupNum}/devices:

    post:
      tags:
        - devices
      summary: Добавляет устройство
      description: Добавляет устройство
      operationId: addDevice
      security:
        - ApiTokenAuth: []
      parameters:
        - name: groupNum
          in: path
          description: Номер группы
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Данные устройства
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      responses:
        '201':
          description: Устройство успешно добавлено
          headers:
            Location:
              description: URL добавленного нового устройства
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceUuid:
                    type: string
                    format: uuid
                    description: Уникальный идентификатор добавленного устройства
          links:
            GetDeviceByUuid:
              operationId: getDevice
              parameters:
                deviceUuid: "$response.body#/deviceUuid"
              description: >-
                Значение `deviceUuid`, полученное в ответе, может использоваться
                в качестве одноимённого параметра в `GET /devices/{deviceUuid}`.
            UpdateDeviceByUuid:
              operationId: updateDevice
              parameters:
                deviceUuid: "$response.body#/deviceUuid"
              description: >-
                Значение `deviceUuid`, полученное в ответе, может использоваться
                в качестве одноимённого параметра в `PUT /devices/{deviceUuid}`.
        '400':
          description: Некорректные данные устройства в запросе или номер группы
        '401':
          description: Несанкционированный доступ
        '404':
          description: Группа не найдена
        default:
          description: Ошибка сервера

    get:
      tags:
        - devices
      summary: Получает устройства группы
      description: Получает устройства группы
      operationId: findAllDevicesByGroup
      security:
        - ApiTokenAuth: []
      parameters:
        - name: groupNum
          in: path
          description: Номер группы
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Данные всех устройств группы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          description: Некорректный номер группы
        '401':
          description: Несанкционированный доступ
        '404':
          description: Группа не найдена
        default:
          description: Ошибка сервера
...
components:
  securitySchemes:
    ApiTokenAuth:
      type: apiKey
      in: header
      name: Api-Token
  schemas:
...
    Group:
      required:
        - name
        - members
      type: object
      properties:
        name:
          type: string
          examples:
            - "Семья"
        members:
          type: array
          items:
            type: object
            properties:
              username:
                type: string
                format: email
          examples:
            - "- username: gertrude@mail.ru
               - username: vasya@inbox.ru
               "
...
    Device:
      required:
        - typeId
        - name
        - modelId
        - serialNum
        - firmwareVersion
        - connectionId
      type: object
      properties:
        deviceUuid:
          type: string
          format: uuid
          examples:
            - 'a83dcf43-dd10-4597-9a9c-6ba1ac50bd0c'
        typeId:
          type: integer
          format: int32
          examples:
            - 1
            - 2
          description: Код производителя типа устройства, например, 1 - робот-пылесос
        name:
          type: string
          examples:
            - Робот Кузя
        modelId:
          type: integer
          format: int32
        serialNum:
          type: string
          examples:
            - "6a0fa287-6969-4839"
        firmwareVersion:
          type: integer
          format: int32
          examples:
            - 2
        connectionId:
          type: integer
          format: int64
          examples:
            - 12
...